Fixed bugs and created more functionality
Created functionality for Switch Card (incomplete), Clear Card, and Back To It Card. Completely finished storage system and its logic.

Overhauled the sequence running system for improved functionality and efficiency.

Overhauled the playing cards system for necessary functionality.

Created Play and Store button so user can select a card, and then choose to either play the card or store the card.

Created two buttons for user to pick to either remove the first card or last card in the sequence when the Clear Card is played.

Added storage cooldown, so the user must wait 2 turns before playing the card they just stored

Created additional functions in DeckManager for user friendly experience and more functionality

Created two more GameStates - SwitchCards and ChooseClear

Disabled store and play buttons if the GameState is not in ChooseCards()

Changed numerous conditions for the GameState to switch automatically